backend/requirements.txt
fastapi>=0.115.5
uvicorn
openai
pydantic
python-dotenv
qdrant-client
pillow
transformers
aiohttp
tqdm
